Separate the kube-toolkit examples and the library
So that we could build a ktk generator command on top of it

Ref #14
