+++ .circleci/config.yml
@@ -28,8 +28,8 @@ jobs:
 
       - run:
           name: make client
-          command: make client
+          command: cd example/demo && make client
 
       - run:
           name: make server
-          command: make server
+          command: cd example/demo && make server
