+++ example/scaffold/Dockerfile
