+++ example/scaffold/Dockerfile.gateway
