+++ example/scaffold/cmd/client/init.go
