+++ example/scaffold/cmd/client/proxy.go
