+++ example/scaffold/cmd/client/serverStream.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"log"
 
-	"github.com/radu-matei/kube-toolkit/pkg/client"
+	"github.com/radu-matei/kube-toolkit/example/scaffold/pkg/client"
 	"github.com/spf13/cobra"
 )
