+++ example/scaffold/cmd/server/server.go
