+++ example/scaffold/pkg/rpc/rpc.pb.gw.go
