+++ example/scaffold/pkg/rpc/rpc.proto
