+++ example/scaffold/pkg/version/version.go
