+++ example/scaffold/scripts/gopath.go
