+++ example/scaffold/scripts/gopkg
